Add error messages and recovery
First, size_t was replaced with a new Location class which contains a
line, column, and absolute position to make errors more human-readable.

For the lexer, it will emit an error if it finds an unknown symbol or an
integer too large to put in a long int.

For the parser, an ErrorNode class was added to implement the Null
Object design pattern instead of throwing a bunch of nullptrs around and
getting a segfault. In order to use it, though, I had to make all the
parsing productions return a Node, rather than a subclass of Node to
take advantage of static type checking.

The lexer and parser both have a special std::ostream& field for
printing errors, initialized to std::cerr by default. This can help in
refactoring if I want to use a different stream for printing out errors.

Add error stream field to lexer