Add semantic verification pass
A NodeVerifier class was added which does all the verification stuff
like symbol tables and type checking.

Token::Type and Node::Type were renamed to Kind so that they don't
conflict with the new class Type (replacing TypeNode) that represents a
VSL type. Secondly, a unique_ptr<Type> field was added for all Nodes for
use by the NodeVerifier.

Prefixes were added to Token::Kind constants to organize it a little bit
better.

'Int' and 'Void' keywords were added for now as the two main primitive
types. There's also an error type but there's no keyword to represent
that in VSL.

Even though I really should be writing unit tests, I added a '-v' option
which starts the verifier REPL.

Fields in Token, Node, and Type were made public because getters are
just redundant and they would've been POD structs anyways if it weren't
for all the polymorphism.

Scope and ScopeTree classes were added for NodeVerifier, where Scope is
like a symbol table and ScopeTree is like a stack of Scopes. It's
slightly more than that, though, since each Scope also stores a return
type so that ReturnNodes can be checked.

Disclaimer: Named parameters are currenetly not supported because of a
limitation in how the function's type represents its own parameters,
which is only storing its type and not also its name.