Fix clang error with dynamic unique_ptrs
A clang bug with c++14 caused he "root" variable in VSLParser::parse()
to not be implicitly downcasted, which is the expected behavior in
c++11. To fix this, I just called make_unique as the return value like
in all the other helper parsers.

Do note however that this is clang 3.8, and that I should probably
update it soon...