Add LLVM IR generator
Most of the basic features that the parser supports are also supported
here, except for maybe constness or other stuff. Keep in mind that this
code isn't really polished at all and will probably blow up in a
segfault if something bad happens for now.

The NodeVerifier class was renamed to LLVMGen and then was extended to
provide one single step to get to LLVM IR. The verifier REPL was
changed accordingly.

Scope and ScopeTree classes now use a Scope::Item class which contains
both a VSL type and an LLVM value so other parts of the LLVMGen can use
it.

The Type class now has a polymorphic toLLVMType() method which
translates a VSL type into an LLVM type that can be used in the IR.