Setup some testing stuff
The testing framework being used is Google Test, which apparently
requires fetching and building from source, preferably through a CMake
ExternalProject, so that took a long time to get it working.

On an unrelated note, randomly renamed the CMake option BUILD_DOCS to
VSL_BUILD_DOCS for consistency with all the other variables.

The LLVM library and the source files except src/main.cpp are put into
a single CMake library target called 'libvsl' so that both the test
program and the compiler itself can have libvsl linked to it without
having to individually link all the object files.

A decent amount of tests for Lexer and Parser were added as well.