Fix obscure error in parsing "-x+1;" as "+x+1;"
The current way that VSLParser::next() was implemented had to do with
removing tokens from the token cache. This didn't work so well with
parseNud(), which calls parseExpr() before it stores the unary operator
minus into the AST. Since the unary operator was stored in a reference,
and that reference was replaced by the next token (the plus), this
caused some undefined behavior with the order of evaluation for
arguments. This also might be why Travis got a segmentation fault using
an older version of gcc and clang while my machine didn't since it had
the newer versions.

Now, instead of how the token cache used to work, it's now a container
that owns all the tokens until the parser is destroyed, guaranteeing
the lifetime of a Token& from current()/next() to be the lifetime of
the actual parser object.

Fixed a typo in the VSLParser documentation.

Removed the peek() method because it wasn't actually being used.

Prevented VSLParser from accepting a unary + operator since LLVMGen
doesn't support it.

Documented most of the parsing productions used in the parser.