Replace main function with Driver class
A new class Driver is supposed to handle all the top level stuff
regarding compilation. It uses an OptionParser class in order to parse
all the arguments and figure out what to do.

The -O<level> option was tweaked because we only have one level for now.

Includes were broken because of the recent change in src/ structure so
that was also fixed.