Reform VSLLexer and Token data
Simplified the Token class and removed all polymorphism. All data was
put under a new TokenKind enum class and an llvm::StringRef field.

Also converted many Node fields to llvm::StringRef instead of
std::string to reduce copying, which was done similarly in Token.

This TokenKind enum class uses a def file so that the entire list of
names can be retrieved multiple times for use in its hpp and cpp files,
similar to how clang does it.

A default constructor was added to Token so that the lexer REPL could
still work without having to use an std::unique_ptr.

Removed the pos field from the Location class since that's already
implied by having a llvm::StringRef field in Token.

Made the lexer REPL include location.

A keyword matcher function was added to make lexing a bit easier, and
its peek() method was simplified to only look one character ahead
because any more is unnecessary for now.

Temporarily fixed a segfault on mismatched int widths. Once more integer
widths and conversions are supported, this will no longer be a problem.