Add Bool while refactoring how ints are handled
A Bool type was kinda added but not fully supported yet. The keywords
Bool, true, and false are there also, but operators are WIP right now.
While they're supported, not all of them make sense to use on a 1bit
integer like addition and multiplication. The rest of the changes below
were what I had to change regarding how ints are handled to allow these
kinds of changes to be more easily made.

NumberExprNode was renamed to IntExprNode since the former didn't
really do anything with floating points.

IntExprNode was changed to hold a llvm::APInt so that it could support
multiple int widths (including 1bit in the case of Bool).

Improved string->int error handling in the parser so that it correctly
detects overflowed and invalid integers.