Add context object
The VSLContext class is supposed to own any long-lived object that has
to be unique, like Type for example. Much of the codebase was refactored
to use const Type* rather than Type&.

Type constructors were made protected so that only subclasses and
friends like VSLContext could access them.

A hasher was added to FunctionType so that VSLContext could store them
in a hash set for easier access.