Make expressions emitted by IRGen more extendable
Separate methods handle each kind of operator, doing all the type
checking independently in each method. This makes it more easily
extendable when adding other operators that don't support every type,
like the logical or/and/not, without having to pile a ton of ifs and
switches ontop of eachother in only a couple methods.