Properly generate if.end blocks and nested ifs
After generating the then/else cases, each would check the last block
it emitted to see if there's some kind of terminator instruction at the
end (like a return for example). If it couldn't find one, it'll create
a branch to the if.end block, causing that block to be included in the
IR. It checks the last block emitted instead of the actual if.then or
if.else block so that it could properly handle nested if statements.