Fix segfault on nested/chained ifs with returns
In an if/else statement, if both cases have a return at the end, then a
merge block won't be created, so the IRBuilder's insertion point would
have to be cleared because there's really nothing to insert after. This
case was handled in a shallow if/else statement when generating the
merge block, but not for a nested or chained if/else with returns, which
causes the segfault when trying to access the IRBuilder's insertion
point which was cleared already.

Also added an error if a function has no return at the end, as well as
another chained if test to make sure stuff is working.