Allow calling functions ahead of their definition
Some **MAJOR** refactoring going on here.

IRGen was turned into an AST pass manager rather than a single AST pass
itself. All of its type checking and LLVM generation code was moved to
an IREmitter pass. Before that, a FuncResolver pass is ran to collect
info on each free function, which the IREmitter uses to generate calls
for functions seemingly before they're defined.

The ScopeTree and Scope classes were completely refactored and split
into global scope (for functions) and function scope (for variables and
parameters within a function body).

NodeVisitor now implements its own virtual functions to do nothing if
they're not overridden. Saves a lot of typing for things like
FuncResolver, which only implements one NodeVisitor method.

An alreadyDefined flag was added to FunctionNode and can be set by
FuncResolver if it believes that the function is trying to be
redefined. IREmitter will ignore it once flagged.

CMake now searches recursively through src instead of just through a
single folder to find the more deeply nested files.

Parser and VSLParser now return a list of statements instead of a
BlockNode to help distinguish between global and function scope.
NodeVisitor was given a method to automatically visit each statement.

VSLContext was moved to the ast folder because its interface only
handles stuff from that folder (types).

A test was added to verify correctness.