Add external functions
Syntax is "func f(x: Int) -> Int external(g);". This allows a function
to exist under one name in VSL, but is secretly referred to as a
different name outside of VSL. This is useful when overloading an
externally-defined function, like for distinguishing between integer
and floating point math functions from <cmath> without having to add
another layer of function calls.