Pretty-print AST instead of using JSON abomination
A visitor called NodePrinter was added which can pretty-print an AST to
a raw_ostream. Normally one wouldnt need to print the AST except when
debugging the parser, so putting all those methods inside Node wouldn't
help very much except in that specific case.

Node::isExpr() was added to help with printing expression statements
with semicolons at the end only when used as a statement and not as,
say, a function call argument.

Type was given a raw_ostream insertion overload so you don't have to
call toString() all the time. Additionally, toString() has been
replaced with a print() method because streams are generally more
efficient than strings that continuously have toreallocate themselves.

The implementation isn't entirely perfect when dealing with unary and
binary operators, though. According to tokenKind.def, operators are set
to be wrapped in single quotes, so `1+2` is incorrectly interpreted as
`1'+'2`. However, since the lexer just emits a warning and skips single
quotes, this shouldn't be a problem if you try to feed the parser's
output back into its input.