Use a separate enum for unary/binary expressions
NodePrinter now doesn't wrap operators with quotes because it uses an
entirely different set of enums.

These enums are UnaryKind and BinaryKind, which specify what operation
a UnaryNode or BinaryNode will calculate.

Improved parser documentation with doxygen grouping as well as some of
the code that has to do with parsing unary and binary expressions.