Add global variables
Added some tests to verify correctness.

GlobalScope's `set()` method was split up into setFunc and setVar to
allow for entering variables into the global scope.

IREmitter was given an llvm::Module& field so that it could do some
module-level stuff like initializing global variables.

IREmitter::visitVariable was almost completely reworked to support
global variables. Some code is still shared between them so they're
still part of the same function.

This uses the `@llvm.global_ctors` intrinsic variable which tells LLVM
to call this `@vsl.init` function at runtime, which will initialize each
global variable in the order they were declared. It will only appear in
the IR if it needs to.

IREmitter::visitIdent was refactored to handle globals since in the LLVM
IR it has a pointer type so that has to be dereferenced first.

What hasn't been done yet is being able to reference global variables
before they show up in the code, similar to how functions work now, as
well as detecting if a variable hasn't been initialized yet for safety.

Replaced the IREmitter::result field with the new Value class. This
uses an enum to determine what kind of value it is (expression,
variable, function, etc.) so that the IREmitter can handle them
differently. Much of irgen/passes and irgen/scope was refactored to use
it instead of VarItem and ExprNode::type.

Some null checks were added all over the place to handle invalid ASTs
while still leaving the generated LLVM IR in a valid state.

In IREmitter's CallNode visitor, the if statement mess was made a bit
tidier with guard clauses.

Fixed a bug in generating relational operators where IREmitter would
treat, say, "1 == 1" as an Int rather than a Bool.

Added a test to verify that some null checks successfully return early,
like trying to call a function that is actually an int variable.