Clearly show when Node pointers can be null
Much of the Node class hierarchy's interface was reworked to incorporate
the philosophy that any Node pointer that isn't stored in a container
may be null. Instead of using a bunch of EmptyNodes to signify a parsing
error, a nullptr is passed. The method that handles it can either use
that nullptr in a Node, if its constructor allows it, or silently fail
and pass the nullptr along.

VSLParser::errorExpected and errorUnexpected no longer return anything
because of this.

The INVALID_PARAM_TYPE diagnostic was changed so that it doesn't require
a ParamNode.