Add reference-counted classes
Lexer changes:
* Keywords added: class, init, and self keywords.

Parser changes:
* Nodes added: ClassNode, FieldNode, CtorNode, MethodNode, SelfNode,
FieldAccessNode, MethodCallNode.
* Types added: ClassType, NamedType.

ClassNodes currently can have multiple fields/methods but only one ctor
until overloading is implemented. The fields that contain these are
mutable so that the ClassNode can be constructed and then its members
which contain references to its parent class.

Added protected constructors to VariableNode and FunctionNode so that
subclasses (fields and methods) can fill in their own Node::Kind.

ClassType stores a hashmap of its fields' name, type, and index within
the class. Name is used for lookup, type for type checking, and index
for the `getelementptr` instruction to access the field.

NamedType represents a type that has a name. The new TypeResolver pass
sets its underlying type field.

FunctionTypes can now represent constructors and methods.

IREmitter changes:

The `self` field was added to keep track of the self
parameter when visiting the body of a method or constructor.

Class types are implemented as a pointer to a struct type (named
`%<class name>`) that contains:
1. An `i32` as a reference count.
2. The actual struct type (named `%struct.<class name>`).

A lot of the code to generate functions were factored out so that
constructors and methods can be done.

Code in IREmitter::visitCall was factored out so that method calls could
be properly done.

Implemented copy and destroy operations where appropriate.

Objects copied in:
* Variable initializers.
* Return values.
* Ternary then/else operands.
* Function/method call arguments.
* Assignment expression rhs.

Objects destroyed in:
* Expression statements inside a block.
* If statements.
* Unary/binary expression operands.
* Ternary conditions.
* Function/method call arguments.
* Method call object.
* Assignment expression rhs.
* Short circuit operands.
* Field access base objects (after copy/store).
* Local variables at end of scope.
* All local variables after return statements and/or end of void
function.
* Global variables at end of program.

Each class will have an auto-generated destructor function that will
decrement the reference count and destroy the object if needed. It takes
into account the fields of the class so that they are destroyed first
before the memory of the object is freed.

Other changes:

VSLContext got a hashmap of NamedTypes and some methods for creating
ClassTypes.

TypeConverter can now convert ClassTypes into its reference-counted
pointer type.

Added field Values which contain info on both the field being accessed
and the base object so that the base can be deleted as soon as the field
is copied/stored. In the case of e.g. `x.y.z`, the generated Value would
contain `z` as the field and `x`, not `y`, as the base object, so that
both `x` and `y` could be destroyed after use.

GlobalScope keeps track of methods assigned to types.

Added a bunch of tests to verify correctness.