Eliminate redundant loads when destroying vars
At end of function scope, all variables have to be destroyed. This was
done by loading the variable Value and calling its destructor if it
exists. By making the "if it exists" part happen before the load, this
can eliminate some of the -O0 IR bloat when a destructor doesn't exist,
no longer leaving leaving a random load instruction at the end.