Allow variables to infer their type from init expr
For now, fields can't have this since inline field inits aren't
implemented yet.

Completely refactored IREmitter::visitVariable to incorporate this
change. Still segfaults when the initializer is omitted because I'm not
sure how to reliably protect from use before initialization.

VSLParser is able to detect if a variable is lacking both an initializer
and a type.

Added a bunch of tests to verify correctness.