Simplify type system
NamedType was removed and is replaced by UnresolvedType. It's very
similar to NamedType except its mutable underlying type field is only
set when needed, in which case it'll look it up from the VSLContext.

Added a bunch of helper methods to Type so we don't have to do
everything in the IREmitter code. This includes checked casting to
derived Types and Type matching/checking.

Type also has a virtual method called resolve which gets the most basic
(resolved) form of the Type, which is useful when looking up ctors/dtors
of a type for example.

ClassType now has its own name field and is stored in VSLContext's
StringMap of other Types that have names. UnresolvedType is not included
here, however, instead being stored in its own container. This is
because they would collide with the names of actual resolved Types.

Since NamedType was replaced, ClassNode::getType() just returns its
ClassType.

VSLContext's interface was refactored to accomodate for all this.

I wasn't exactly sure when and where to call Type's resolve and match
methods so I threw it around a lot in IREmitter. Later I think I'll try
to take advantage of the type system to make it less error-prone.

TypeResolver now leaves all of the initial creation of LLVM ClassType
equivalent objects to the TypeConverter.